webkit/data/layout_tests/chrome/fast/dom/dom-constructors-expected.txt

This test checks that all but a handful of dom constructors throw exceptions, and the rest return reasonable objects

Attr: exception
CharacterData: exception
CDATASection: exception
Comment: exception
Document: exception
DocumentFragment: exception
DocumentType: exception
Element: exception
Entity: exception
EntityReference: exception
EventTargetNode: no constructor
HTMLDocument: exception
Node: exception
Notation: exception
ProcessingInstruction: exception
Text: exception
HTMLAnchorElement: exception
HTMLAppletElement: exception
HTMLAreaElement: exception
HTMLBaseElement: exception
HTMLBaseFontElement: exception
HTMLBlockquoteElement: exception
HTMLBodyElement: exception
HTMLBRElement: exception
HTMLButtonElement: exception
HTMLCanvasElement: exception
HTMLDirectoryElement: exception
HTMLDivElement: exception
HTMLDListElement: exception
HTMLEmbedElement: exception
HTMLFieldSetElement: exception
HTMLFontElement: exception
HTMLFormElement: exception
HTMLFrameElement: exception
HTMLFrameSetElement: exception
HTMLHeadingElement: exception
HTMLHeadElement: exception
HTMLHRElement: exception
HTMLHtmlElement: exception
HTMLIFrameElement: exception
HTMLImageElement: exception
HTMLInputElement: exception
HTMLIsIndexElement: exception
HTMLLabelElement: exception
HTMLLegendElement: exception
HTMLLIElement: exception
HTMLLinkElement: exception
HTMLMapElement: exception
HTMLMarqueeElement: exception
HTMLMenuElement: exception
HTMLMetaElement: exception
HTMLModElement: exception
HTMLObjectElement: exception
HTMLOListElement: exception
HTMLOptGroupElement: exception
HTMLOptionElement: exception
HTMLParagraphElement: exception
HTMLParamElement: exception
HTMLPreElement: exception
HTMLQuoteElement: exception
HTMLScriptElement: exception
HTMLSelectElement: exception
HTMLStyleElement: exception
HTMLTableCaptionElement: exception
HTMLTableColElement: exception
HTMLTableElement: exception
HTMLTableSectionElement: exception
HTMLTableCellElement: exception
HTMLTableRowElement: exception
HTMLTextAreaElement: exception
HTMLTitleElement: exception
HTMLUListElement: exception
HTMLElement: exception
BarInfo: no constructor
CanvasGradient: no constructor
CanvasPattern: no constructor
CanvasRenderingContext2D: exception
Clipboard: exception
Console: no constructor
Counter: exception
CSSCharsetRule: exception
CSSFontFaceRule: exception
CSSImportRule: exception
CSSMediaRule: exception
CSSPageRule: exception
CSSPrimitiveValue: exception
CSSRule: exception
CSSRuleList: exception
CSSStyleDeclaration: exception
CSSStyleRule: exception
CSSStyleSheet: exception
CSSValue: exception
CSSValueList: exception
DOMImplementation: exception
DOMParser: [object DOMParser]
DOMSelection: no constructor
DOMWindow: no constructor
Event: exception
History: no constructor
UndetectableHTMLCollection: no constructor
HTMLCollection: exception
HTMLOptionsCollection: no constructor
InspectorController: no constructor
KeyboardEvent: exception
Location: no constructor
MediaList: exception
MimeType: exception
MimeTypeArray: exception
MouseEvent: exception
MutationEvent: exception
NamedNodeMap: exception
Navigator: no constructor
NodeFilter: exception
NodeIterator: no constructor
NodeList: exception
OverflowEvent: exception
Plugin: exception
PluginArray: exception
Range: exception
Rect: exception
RGBColor: no constructor
Screen: no constructor
StyleSheet: exception
StyleSheetList: exception
TextEvent: exception
TreeWalker: no constructor
UIEvent: exception
WheelEvent: exception
XMLHttpRequest: [object XMLHttpRequest]
XMLSerializer: [object XMLSerializer]
XPathEvaluator: [object XPathEvaluator]
XPathExpression: no constructor
XPathNSResolver: no constructor
XPathResult: exception
XSLTProcessor: [object XSLTProcessor]
EventTarget: no constructor
EventListener: no constructor
NPObject: no constructor
Console: no constructor
